move devlivery classes into new namespace
