+++ Kentico.Kontent.Delivery.Caching.Tests/CacheHelpersTests.cs
@@ -2,8 +2,8 @@
 using FluentAssertions;
 using Kentico.Kontent.Delivery.Abstractions;
 using Kentico.Kontent.Delivery.Caching.Tests.ContentTypes;
-using Kentico.Kontent.Urls.QueryParameters;
-using Kentico.Kontent.Urls.QueryParameters.Filters;
+using Kentico.Kontent.Urls.Delivery.QueryParameters;
+using Kentico.Kontent.Urls.Delivery.QueryParameters.Filters;
 using Xunit;
 
 namespace Kentico.Kontent.Delivery.Caching.Tests
