+++ Kentico.Kontent.Delivery.Rx.Tests/DeliveryObservableProxyTests.cs
@@ -10,8 +10,8 @@ using FakeItEasy;
 using Kentico.Kontent.Delivery.Abstractions;
 using Kentico.Kontent.Delivery.ContentItems;
 using Kentico.Kontent.Delivery.Rx.Tests.Models.ContentTypes;
-using Kentico.Kontent.Urls.QueryParameters;
-using Kentico.Kontent.Urls.QueryParameters.Filters;
+using Kentico.Kontent.Urls.Delivery.QueryParameters;
+using Kentico.Kontent.Urls.Delivery.QueryParameters.Filters;
 using Microsoft.Extensions.Options;
 using RichardSzalay.MockHttp;
 using Xunit;
