+++ Kentico.Kontent.Delivery.Tests/DeliveryClientTests.cs
@@ -14,8 +14,8 @@ using Kentico.Kontent.Delivery.ContentItems.RichText.Blocks;
 using Kentico.Kontent.Delivery.Tests.Factories;
 using Kentico.Kontent.Delivery.Tests.Models;
 using Kentico.Kontent.Delivery.Tests.Models.ContentTypes;
-using Kentico.Kontent.Urls.QueryParameters;
-using Kentico.Kontent.Urls.QueryParameters.Filters;
+using Kentico.Kontent.Urls.Delivery.QueryParameters;
+using Kentico.Kontent.Urls.Delivery.QueryParameters.Filters;
 using Newtonsoft.Json.Linq;
 using RichardSzalay.MockHttp;
 using Xunit;
