+++ Kentico.Kontent.Delivery.Tests/QueryParameters/ContentTypeExtractorTests.cs
@@ -4,8 +4,8 @@ using FakeItEasy;
 using Kentico.Kontent.Delivery.Abstractions;
 using Kentico.Kontent.Delivery.ContentItems;
 using Kentico.Kontent.Delivery.Tests.Factories;
-using Kentico.Kontent.Urls.QueryParameters;
-using Kentico.Kontent.Urls.QueryParameters.Filters;
+using Kentico.Kontent.Urls.Delivery.QueryParameters;
+using Kentico.Kontent.Urls.Delivery.QueryParameters.Filters;
 using Xunit;
 
 namespace Kentico.Kontent.Delivery.Tests.QueryParameters
