+++ Kentico.Kontent.Delivery/DeliveryClient.cs
@@ -12,9 +12,9 @@ using Kentico.Kontent.Delivery.Extensions;
 using Kentico.Kontent.Delivery.Languages;
 using Kentico.Kontent.Delivery.SharedModels;
 using Kentico.Kontent.Delivery.TaxonomyGroups;
-using Kentico.Kontent.Urls;
-using Kentico.Kontent.Urls.QueryParameters;
-using Kentico.Kontent.Urls.QueryParameters.Filters;
+using Kentico.Kontent.Urls.Delivery;
+using Kentico.Kontent.Urls.Delivery.QueryParameters;
+using Kentico.Kontent.Urls.Delivery.QueryParameters.Filters;
 using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
