+++ Kentico.Kontent.Urls/Delivery/QueryParameters/DepthParameter.cs
@@ -1,6 +1,6 @@
 ﻿using Kentico.Kontent.Delivery.Abstractions;
 
-namespace Kentico.Kontent.Urls.QueryParameters
+namespace Kentico.Kontent.Urls.Delivery.QueryParameters
 {
     /// <summary>
     /// Specifies the maximum level of recursion when retrieving linked items. If not specified, the default depth is one level.
