+++ Kentico.Kontent.Urls/Delivery/QueryParameters/Filters/Filter.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using Kentico.Kontent.Delivery.Abstractions;
 
-namespace Kentico.Kontent.Urls.QueryParameters.Filters
+namespace Kentico.Kontent.Urls.Delivery.QueryParameters.Filters
 {
     /// <summary>
     /// Provides the base class for filter implementations.
