+++ Kentico.Kontent.Urls/Delivery/QueryParameters/OrderParameter.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Kentico.Kontent.Delivery.Abstractions;
 
-namespace Kentico.Kontent.Urls.QueryParameters
+namespace Kentico.Kontent.Urls.Delivery.QueryParameters
 {
     /// <summary>
     /// Specifies how content items are sorted.
